requirements.txt
streamlit==1.51.0
openai==2.6.1
python-dotenv==1.2.1
pip==25.3